Move write methods to project, branch filtering to own file
