+++ project.go
@@ -1,15 +1,14 @@
 package main
 
 import (
-	"bufio"
 	"bytes"
 	"fmt"
+	"html/template"
 	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"strings"
-	"time"
 )
 
 // SEP is a browser generated UUID v4 used to separate out commit line items.
@@ -20,20 +19,44 @@ var types = make(map[string]bool)
 
 type project struct {
 	base     string
-	Branches []branch
 	Name     string
 	repo     string
+	options  *options
+	template *template.Template
+	branches []branch
+}
+
+func NewProject(base string, name string, repo string, options *options) *project {
+	funcMap := template.FuncMap{
+		"diffstatbodyparser": diffstatbodyparser,
+		"diffbodyparser":     diffbodyparser,
+	}
+	template := template.Must(template.New("page").Funcs(funcMap).Parse(tpl))
+
+	branches, err := branchFilter(repo, name, options)
+	if err != nil {
+		log.Fatalf("unable to filter branches: %v", err)
+	}
+
+	return &project{
+		base:     base,
+		Name:     name,
+		repo:     repo,
+		options:  options,
+		template: template,
+		branches: branches,
+	}
 }
 
 // Creates base directories for holding objects, branches, and commits.
-func (pro *project) init(f bool) error {
+func (p *project) init() error {
 	dirs := []string{"branch", "commit", "object"}
 
 	for _, dir := range dirs {
-		d := filepath.Join(pro.base, dir)
+		d := filepath.Join(p.base, dir)
 
 		// Clear existing dirs when -f true.
-		if f && dir != "branch" {
+		if p.options.Force && dir != "branch" {
 			if err := os.RemoveAll(d); err != nil {
 				return fmt.Errorf("unable to remove directory: %v", err)
 			}
@@ -48,230 +71,305 @@ func (pro *project) init(f bool) error {
 }
 
 // Saves a local clone of `target` repo.
-func (pro *project) save(target string) error {
-	if _, err := os.Stat(pro.repo); err != nil {
+func (p *project) save() error {
+	if _, err := os.Stat(p.repo); err != nil {
 		return err
 	}
 
-	return exec.Command("git", "clone", target, pro.repo).Run()
+	return exec.Command("git", "clone", p.options.Source, p.repo).Run()
 }
 
-// Goes through list of branches and returns those that match whitelist.
-func (pro *project) branchfilter(whitelist manyflag) ([]branch, error) {
-	cmd := exec.Command("git", "branch", "-a")
-	cmd.Dir = pro.repo
+func (p *project) updateBranches() {
+	for _, b := range p.branches {
+		// NOTE: Is this needed still if the repo is downloaded each time the script is run?
+		ref := fmt.Sprintf("refs/heads/%s:refs/origin/%s", b, b)
+
+		cmd := exec.Command("git", "fetch", "--force", "origin", ref)
+		cmd.Dir = p.repo
+
+		log.Printf("updating branch: %s", b)
+
+		if _, err := cmd.Output(); err != nil {
+			log.Printf("unable to fetch branch: %v", err)
+			continue
+		}
+	}
+}
+
+func (p *project) writePages() {
+	for _, b := range p.branches {
+		log.Printf("processing branch: %s", b)
+
+		go p.writeBranchPage(b)
+
+		for i, c := range b.Commits {
+			log.Printf("processing commit: %s: %d/%d", c.Abbr, i+1, len(b.Commits))
+
+			base := filepath.Join(p.base, "commit", c.Hash)
+
+			if err := os.MkdirAll(base, 0755); err != nil {
+				if err != nil {
+					log.Printf("unable to create commit directory: %v", err)
+				}
+
+				continue
+			}
+
+			for _, par := range c.Parents {
+				p.writeCommitDiff(par, c, base, b)
+			}
+
+			for _, obj := range c.Tree {
+				dst := filepath.Join(p.base, "object", obj.Dir())
+
+				if err := os.MkdirAll(filepath.Dir(dst), 0755); err != nil {
+					if err != nil {
+						log.Printf("unable to create object directory: %v", err)
+					}
+					continue
+				}
+
+				p.writeObjectBlob(obj, dst)
+				p.writeNom(fmt.Sprintf("%s.html", dst), obj, b, c, base)
+			}
+
+			p.writeCommitPage(base, c, b)
+		}
+	}
+}
+
+func (p *project) writeMainIndex() {
+	// This is the main index or project home.
+	f, err := os.Create(filepath.Join(p.base, "index.html"))
+
+	defer f.Close()
+
+	if err != nil {
+		log.Fatalf("unable to create home page: %v", err)
+	}
+
+	page := page{
+		Data: Data{
+			"Branches": p.branches,
+			"Link":     p.options.URL,
+			"Project":  p.Name,
+		},
+		Base:  "./",
+		Title: p.Name,
+	}
+
+	if err := p.template.Execute(f, page); err != nil {
+		log.Fatalf("unable to apply template: %v", err)
+	}
+}
+
+func (p *project) writeCommitDiff(par string, c commit, base string, b branch) {
+	cmd := exec.Command("git", "diff", "-p", fmt.Sprintf("%s..%s", par, c.Hash))
+	cmd.Dir = p.repo
 
 	out, err := cmd.Output()
 
 	if err != nil {
-		return nil, err
+		log.Printf("unable to diff against parent: %v", err)
+
+		return
 	}
 
-	var b = make(map[string]branch)
-	var m = make(map[string]bool)
+	dst := filepath.Join(base, fmt.Sprintf("diff-%s.html", par))
+	f, err := os.Create(dst)
 
-	scanner := bufio.NewScanner(bytes.NewReader(out))
+	defer f.Close()
 
-	for scanner.Scan() {
-		t := strings.TrimSpace(strings.TrimPrefix(scanner.Text(), "*"))
-		_, f := filepath.Split(t)
+	if err != nil {
+		log.Printf("unable to create commit diff to parent: %v", err)
 
-		m[f] = true
+		return
 	}
 
-	if err := scanner.Err(); err != nil {
-		return nil, err
+	page := page{
+		Data: Data{
+			"Diff": diff{
+				Body:   fmt.Sprintf("%s", out),
+				Commit: c,
+				Parent: par,
+			},
+			"Project": p.Name,
+		},
+		Base:  "../../",
+		Title: strings.Join([]string{p.Name, b.Name, c.Abbr}, ": "),
 	}
 
-	// Filter to match options, but return all if no branch flags given.
-	if len(whitelist) > 0 {
-		for k := range m {
-			m[k] = contains(whitelist, k)
+	if err := p.template.Execute(f, page); err != nil {
+		log.Printf("unable to apply template: %v", err)
+
+		return
+	}
+}
+
+func (p *project) writeBranchPage(b branch) {
+	dst := filepath.Join(p.base, "branch", b.Name, "index.html")
+
+	if err := os.MkdirAll(filepath.Dir(dst), 0755); err != nil {
+		if err != nil {
+			log.Fatalf("unable to create branch directory: %v", err)
 		}
+		return
+	}
+
+	f, err := os.Create(dst)
+
+	defer f.Close()
+
+	if err != nil {
+		// TODO: Remove from branches slice?
+		log.Printf("unable to create branch page: %v", err)
+
+		return
+	}
+
+	page := page{
+		Data: Data{
+			"Commits": b.Commits,
+			"Branch":  b,
+			"Project": p.Name,
+		},
+		Base:  "../../",
+		Title: strings.Join([]string{p.Name, b.Name}, ": "),
+	}
+
+	if err := p.template.Execute(f, page); err != nil {
+		log.Printf("unable to apply template: %v", err)
+		return
+	}
+}
+
+func (p *project) writeObjectBlob(obj object, dst string) {
+	cmd := exec.Command("git", "cat-file", "blob", obj.Hash)
+	cmd.Dir = p.repo
+
+	out, err := cmd.Output()
+
+	if err != nil {
+		log.Printf("unable to save object: %v", err)
+		return
+	}
+
+	f, err := os.Create(dst)
+
+	defer f.Close()
+
+	if err != nil {
+		log.Printf("unable to create object: %v", err)
+		return
+	}
+
+	if _, err := f.Write(out); err != nil {
+		log.Printf("unable to write object blob: %v", err)
+		return
+	}
+}
+
+func (p *project) writeNom(nom string, obj object, b branch, c commit, base string) {
+	f, err := os.Create(nom)
+	defer f.Close()
+
+	if err != nil {
+		log.Printf("unable to create object: %v", err)
+		return
+	}
+
+	o := &show{
+		object: object{
+			Hash: obj.Hash,
+			Path: obj.Path,
+		},
+		Bin: types[filepath.Ext(obj.Path)],
+	}
+
+	if o.Bin {
+		// TODO.
 	} else {
-		// In git given order at this point.
-		for k := range m {
-			whitelist = append(whitelist, k)
-		}
-	}
+		cmd := exec.Command("git", "show", "--no-notes", obj.Hash)
+		cmd.Dir = p.repo
 
-	for k, v := range m {
-		if v {
-			// TODO: Try a goroutine?
-			commits, err := pro.commitparser(k)
-
-			if err != nil {
-				continue
-			}
-
-			b[k] = branch{commits, k, pro.Name}
-		}
-	}
-
-	// Fill in resulting slice with desired branches in order.
-	var results []branch
-
-	for _, v := range whitelist {
-		results = append(results, b[v])
-	}
-
-	return results, nil
-}
-
-func (pro *project) commitparser(b string) ([]commit, error) {
-	fst := strings.Join([]string{"%H", "%P", "%s", "%aN", "%aE", "%aD", "%h"}, SEP)
-	ref := fmt.Sprintf("origin/%s", b)
-
-	cmd := exec.Command("git", "log", fmt.Sprintf("--format=%s", fst), ref)
-	cmd.Dir = pro.repo
-
-	out, err := cmd.Output()
-
-	if err != nil {
-		return nil, err
-	}
-
-	results := []commit{}
-	scanner := bufio.NewScanner(bytes.NewReader(out))
-
-	for scanner.Scan() {
-		text := strings.TrimSpace(scanner.Text())
-		data := strings.Split(text, SEP)
-
-		h := data[0]
-
-		var history []overview
-		var parents []string
-
-		if data[1] != "" {
-			parents = strings.Split(data[1], " ")
-		}
-
-		for _, p := range parents {
-			diffstat, err := pro.diffstatparser(h, p)
-
-			if err != nil {
-				log.Printf("unable to diffstat against parent: %s", err)
-
-				continue
-			}
-
-			history = append(history, overview{diffstat, h, p})
-		}
-
-		a := author{data[4], data[3]}
-
-		date, err := time.Parse("Mon, 2 Jan 2006 15:04:05 -0700", data[5])
+		out, err := cmd.Output()
 
 		if err != nil {
-			log.Printf("unable to parse commit date: %s", err)
+			log.Printf("unable to show object: %v", err)
 
-			continue
+			return
 		}
 
-		body, err := pro.bodyparser(h)
+		sep := []byte("\n")
+		var lines = make([]int, bytes.Count(out, sep))
 
+		for i := range lines {
+			lines[i] = i + 1
+		}
+
+		if bytes.LastIndex(out, sep) != len(out)-1 {
+			lines = append(lines, len(lines))
+		}
+
+		o.Lines = lines
+		o.Body = fmt.Sprintf("%s", out)
+	}
+
+	page := page{
+		Data: Data{
+			"Object":  *o,
+			"Project": p.Name,
+		},
+		Base:  "../../",
+		Title: strings.Join([]string{p.Name, b.Name, c.Abbr, obj.Path}, ": "),
+	}
+
+	if err := p.template.Execute(f, page); err != nil {
+		log.Printf("unable to apply template: %v", err)
+		return
+	}
+
+	lnk := filepath.Join(base, fmt.Sprintf("%s.html", obj.Path))
+
+	if err := os.MkdirAll(filepath.Dir(lnk), 0755); err != nil {
 		if err != nil {
-			log.Printf("unable to parse commit body: %s", err)
-
-			continue
+			log.Printf("unable to create hard link path: %v", err)
 		}
-
-		tree, err := pro.treeparser(h)
-
-		if err != nil {
-			log.Printf("unable to parse commit tree: %s", err)
-
-			continue
-		}
-
-		c := commit{
-			Abbr:    data[6],
-			Author:  a,
-			Body:    body,
-			Branch:  b,
-			Date:    date,
-			Hash:    h,
-			History: history,
-			Parents: parents,
-			Project: pro.Name,
-			Subject: data[2],
-			Tree:    tree,
-		}
-
-		results = append(results, c)
+		return
 	}
 
-	if err := scanner.Err(); err != nil {
-		return nil, err
-	}
+	if err := os.Link(nom, lnk); err != nil {
+		if os.IsExist(err) {
+			return
+		}
 
-	return results, nil
+		log.Printf("unable to hard link object into commit folder: %v", err)
+	}
 }
 
-func (pro *project) treeparser(h string) ([]object, error) {
-	cmd := exec.Command("git", "ls-tree", "-r", "--format=%(objectname) %(path)", h)
-	cmd.Dir = pro.repo
+func (p *project) writeCommitPage(base string, c commit, b branch) {
+	dst := filepath.Join(base, "index.html")
+	f, err := os.Create(dst)
 
-	out, err := cmd.Output()
-
-	if err != nil {
-		return nil, err
-	}
-
-	var results []object
-	feed := strings.Split(strings.TrimSuffix(fmt.Sprintf("%s", out), "\n"), "\n")
-
-	for _, line := range feed {
-		w := strings.Split(line, " ")
-
-		results = append(results, object{
-			Hash: w[0],
-			Path: w[1],
-		})
-	}
-
-	return results, nil
-}
-
-func (pro *project) diffstatparser(h, p string) (string, error) {
-	cmd := exec.Command("git", "diff", "--stat", fmt.Sprintf("%s..%s", p, h))
-	cmd.Dir = pro.repo
-
-	out, err := cmd.Output()
+	defer f.Close()
 
 	if err != nil {
-		return "", err
+		log.Printf("unable to create commit page: %v", err)
+		// TODO(spike): handle error?
+		return
 	}
 
-	var results []string
-	feed := strings.Split(strings.TrimSuffix(fmt.Sprintf("%s", out), "\n"), "\n")
-
-	for _, line := range feed {
-		// NOTE: This is hackish I know, attach to project?
-		i := strings.Index(line, "|")
-
-		if i != -1 {
-			ext := filepath.Ext(strings.TrimSpace(line[:i]))
-			types[ext] = strings.Contains(line, "Bin")
-		}
-
-		results = append(results, strings.TrimSpace(line))
+	page := page{
+		Data: Data{
+			"Commit":  c,
+			"Project": p.Name,
+		},
+		Base:  "../../",
+		Title: strings.Join([]string{p.Name, b.Name, c.Abbr}, ": "),
 	}
 
-	return strings.Join(results, "\n"), nil
-}
-
-func (pro *project) bodyparser(h string) (string, error) {
-	// Because the commit message body is multiline and is tripping the scanner.
-	cmd := exec.Command("git", "show", "--no-patch", "--format=%B", h)
-	cmd.Dir = pro.repo
-
-	out, err := cmd.Output()
-
-	if err != nil {
-		return "", err
+	if err := p.template.Execute(f, page); err != nil {
+		log.Printf("unable to apply template: %v", err)
+		return
 	}
-
-	return strings.TrimSuffix(fmt.Sprintf("%s", out), "\n"), nil
 }
